requirements.txt
streamlit>=1.30.0
google-generativeai>=0.8.3
pytz
duckduckgo-search>=5.0.0
# final_stable_v5